+++ app/src/main/java/com/primoz/timetimer/data_mvp/WorkoutsRepository.kt
@@ -2,17 +2,11 @@ package com.primoz.timetimer.data_mvp
 
 import io.realm.RealmList
 
-
-/**
- * Created by Primož Pesjak on 22/09/2017.
- */
-
 class WorkoutsRepository(private val mWorkoutsDataSource: WorkoutsDataSource) : WorkoutsDataSource {
 
     fun destroyInstance() {
         INSTANCE = null
     }
-
 
     override fun getWorkouts(callback: WorkoutsDataSource.LoadWorkoutsCallback) {
         mWorkoutsDataSource.getWorkouts(object : WorkoutsDataSource.LoadWorkoutsCallback {
@@ -27,7 +21,7 @@ class WorkoutsRepository(private val mWorkoutsDataSource: WorkoutsDataSource) : 
         })
     }
 
-    override fun getWorkout(workoutID: Long, callback: WorkoutsDataSource.GetWorkoutCallback) {
+    override fun getWorkout(workoutID: Int, callback: WorkoutsDataSource.GetWorkoutCallback) {
         mWorkoutsDataSource.getWorkout(workoutID, callback)
     }
 
@@ -35,11 +29,11 @@ class WorkoutsRepository(private val mWorkoutsDataSource: WorkoutsDataSource) : 
         mWorkoutsDataSource.saveWorkout(workout, callback)
     }
 
-    override fun editWorkout(workoutID: Long, workout: WorkoutPOJO, callback: WorkoutsDataSource.EditedWorkoutCallback) {
+    override fun editWorkout(workoutID: Int, workout: WorkoutPOJO, callback: WorkoutsDataSource.EditedWorkoutCallback) {
         mWorkoutsDataSource.editWorkout(workoutID, workout, callback)
     }
 
-    override fun deleteWorkout(workoutID: Long, callback: WorkoutsDataSource.DeletedWorkoutCallback) {
+    override fun deleteWorkout(workoutID: Int, callback: WorkoutsDataSource.DeletedWorkoutCallback) {
         mWorkoutsDataSource.deleteWorkout(workoutID, callback)
     }
